requirements.txt
streamlit
cohere>=5.1.8,<5.2.0
langchain==0.1.16
langchain-core==0.1.53
langchain-community==0.0.32
langchain-cohere==0.1.1
PyPDF2
faiss-cpu
python-dotenv
